move eval into eval.rs
