+++ src/core/ir/mod.rs
@@ -10,10 +10,6 @@ mod resolver_context_like;
 
 use std::collections::HashMap;
 use std::fmt::Debug;
-use std::future::Future;
-use std::ops::Deref;
-
-use async_graphql_value::ConstValue;
 pub use cache::*;
 pub use error::*;
 pub use eval::*;
@@ -24,8 +20,7 @@ pub use resolver_context_like::{EmptyResolverContext, ResolverContext, ResolverC
 use strum_macros::Display;
 
 use crate::core::blueprint::DynamicValue;
-use crate::core::json::JsonLike;
-use crate::core::serde_value_ext::ValueExt;
+
 
 #[derive(Clone, Debug, Display)]
 pub enum IR {
@@ -62,92 +57,4 @@ pub struct Map {
     pub input: Box<IR>,
     // accept key return value instead of
     pub map: HashMap<String, String>,
-}
-
-impl Eval for IR {
-    #[tracing::instrument(skip_all, fields(otel.name = %self), err)]
-    fn eval<Ctx>(
-        &self,
-        ctx: &mut EvaluationContext<'_, Ctx>,
-    ) -> impl Future<Output = Result<ConstValue, EvaluationError>>
-    where
-        Ctx: ResolverContextLike + Sync,
-    {
-        Box::pin(async move {
-            match self {
-                IR::Context(op) => match op {
-                    Context::Value => {
-                        Ok(ctx.value().cloned().unwrap_or(async_graphql::Value::Null))
-                    }
-                    Context::Path(path) => Ok(ctx
-                        .path_value(path)
-                        .map(|a| a.into_owned())
-                        .unwrap_or(async_graphql::Value::Null)),
-                    Context::PushArgs { expr, and_then } => {
-                        let args = expr.eval(ctx).await?;
-                        let ctx = &mut ctx.with_args(args);
-                        and_then.eval(ctx).await
-                    }
-                    Context::PushValue { expr, and_then } => {
-                        let value = expr.eval(ctx).await?;
-                        let ctx = &mut ctx.with_value(value);
-                        and_then.eval(ctx).await
-                    }
-                },
-                IR::Path(input, path) => {
-                    let inp = &input.eval(ctx).await?;
-                    Ok(inp
-                        .get_path(path)
-                        .unwrap_or(&async_graphql::Value::Null)
-                        .clone())
-                }
-                IR::Dynamic(value) => Ok(value.render_value(ctx)),
-                IR::Protect(expr) => {
-                    ctx.request_ctx
-                        .auth_ctx
-                        .validate(ctx.request_ctx)
-                        .await
-                        .to_result()?;
-                    expr.eval(ctx).await
-                }
-                IR::IO(operation) => operation.execute(ctx).await,
-                IR::Cache(Cache { max_age, io }) => {
-                    let expr = io.deref();
-                    let key = io.cache_key(ctx);
-                    if let Some(key) = key {
-                        if let Some(val) = ctx.request_ctx.runtime.cache.get(&key).await? {
-                            Ok(val)
-                        } else {
-                            let val = expr.execute(ctx).await?;
-                            ctx.request_ctx
-                                .runtime
-                                .cache
-                                .set(key, val.clone(), max_age.to_owned())
-                                .await?;
-                            Ok(val)
-                        }
-                    } else {
-                        expr.execute(ctx).await
-                    }
-                }
-                IR::Map(Map { input, map }) => {
-                    let value = input.eval(ctx).await?;
-                    if let ConstValue::String(key) = value {
-                        if let Some(value) = map.get(&key) {
-                            Ok(ConstValue::String(value.to_owned()))
-                        } else {
-                            Err(EvaluationError::ExprEvalError(format!(
-                                "Can't find mapped key: {}.",
-                                key
-                            )))
-                        }
-                    } else {
-                        Err(EvaluationError::ExprEvalError(
-                            "Mapped key must be string value.".to_owned(),
-                        ))
-                    }
-                }
-            }
-        })
-    }
 }
